.github: Use checkout@v3 for Bazel
The other jobs were swapped to v3 in 38311e8730. Bazel was added later
with the older version.

This fixes the warning:
```
The following actions uses node12 which is deprecated and will be forced to run on node16: actions/checkout@v2. For more info: https://github.blog/changelog/2023-06-13-github-actions-all-actions-will-run-on-node16-instead-of-node12-by-default/
```