bazel: Don't use artifact() for protobuf deps
We don't include protobuf in IO_GRPC_GRPC_JAVA_ARTIFACTS, so there might
not actually be an alias available for it to @com_google_protobuf. While
we could add it, it is easier to use the @com_google_protobuf references
directly.

This was preventing `bazel query 'deps(//...)' from succeeding, because
it couldn't find javalite.